daemon/config: inform the user when the configuration is not valid UTF-8
This is a minimal approach; without a richer UTF-8 validator we cannot
inform the user at what byte offset the error occurred, or guess what
what the encoding might be.

However, we can reduce user confusion due to otherwise
inscrutable-to-the-uninitiated errors like the following:

> invalid character 'þ' looking for beginning of value
> invalid character 'ÿ' looking for beginning of value

While meaningful to those who are familiar with and are equipped to
diagnose encoding issues, these characters will be hidden when the file
is edited with a BOM-aware text editor, and further confuse the user.

Signed-off-by: Bjorn Neergaard <bneergaard@mirantis.com>