Revert "libnetwork: Support IPv6 in arrangeUserFilterRule()"
This reverts commit 2d397beb0068ebf42f7196e243f33842544693d5.

moby#44706 and moby#44805 were both merged, and both refactored the same
file. The combination broke the build, and was not detected in CI as
only the combination of the two, applied to the same parent commit,
caused the failure.

moby#44706 should be carried forward, based on the current master, in
order to resolve this conflict.

Signed-off-by: Bjorn Neergaard <bneergaard@mirantis.com>