Fix TLS configuration resolution for buildkit
The resolver in `daemon.(*Daemon).RegistryHosts()` returned
`docker.RegistryHost` structs with clients that were not able to
authenticate with registries requiring client TLS certificates.

This patch adds a unit test for that functionality, verifying that
the `http.Client` in the returned `docker.RegistryHost` is able to
perform a TLS Handshake with a mutually-authenticated TLS registry
having a corresponding host certificate directory in a manner
consistent with the actual behavior of the `docker image pull`
command and the documented behavior for verifying repository clients
with certificates [1].

Additionally, a fix is included in `daemon/daemon.go`, replacing the
lines that previously offloaded TLS configuration resolution to the
buildkit package for all registries having associated TLS configuration
directories with logic that uses the `registry.ReadCertsDirectory`
method used elsewhere in this repository for certificate resolution.

It appears to my reading that there is some subtle difference between
the TLS configuration resolution in the `registry` package of this
repository and that in the `util/resolver` package of the buildkit
repository, and that does still leave me with some concerns because the
solution I have implemented does not cover every case:
TLS configurations for explicitly-declared insecure registries and
mirrors of `docker.io` are still resolved by buildkit's resolver, which
demonstrably did not work for the general case of secure registries
covered by this patch. I suspect there may still be a bug lurking in
Buildkit's implementation of the TLS configuration resolution behavior.

1: https://docs.docker.com/engine/security/certificates/

Resolves #38386

Signed-off-by: Peter <peter@psanders.me>