Use streams with container attach
This allows a registered stream to be used with container attach.
There is no multiplexing, connection hijacking, or other weirdness.
This is a raw attachment.

If someone decides to use the same stream id for multiple attachments
(for the same container or multiple containers) it will happily oblige
and the data will be merged as one stream as one might expect if you
have multiple writers to a single pipe.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>