Set BuildKit version and revision at build time
We should set the BuildKit version and revision at build time so
we make sure we are aligned with the expected vendoring.

Also needs to remove the dependency on git in with-go-mod script as we
ignore .git folder in dockerignore, otherwise it fails. And needs to
make bind mount writable for the go.mod file as well.

Signed-off-by: CrazyMax <crazy-max@users.noreply.github.com>