Add workflow to automatically cherry-pick on merge
When a PR is merged that has a cherry-pick label such as
`process/cherry-pick/<branch>`, this workflow will kick in and attempt
to open a backport PR.
If the commits do not apply cleanly it will just error out.

This requires a settings change to allow github actions to approve and
open PR's.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>